project/requirements.txt
tinyfish>=0.2.4
python-dotenv>=1.0.0
openai>=1.40.0
